remove type param from type lattice
