+++ src/kirin/dialects/py/types/__init__.py
@@ -1,18 +1,4 @@
 from kirin.dialects.py.types import emit as emit
-from kirin.dialects.py.types.base import (
-    PyAnyType as PyAnyType,
-    PyBottomType as PyBottomType,
-    PyClass as PyClass,
-    PyConst as PyConst,
-    PyGeneric as PyGeneric,
-    PyLiteral as PyLiteral,
-    PyType as PyType,
-    PyTypeVar as PyTypeVar,
-    PyUnion as PyUnion,
-    PyVararg as PyVararg,
-    hint2type as hint2type,
-    widen_const as widen_const,
-)
 from kirin.dialects.py.types.builtin import (
     Any as Any,
     Bool as Bool,
@@ -31,3 +17,18 @@ from kirin.dialects.py.types.builtin import (
     TypeofFunctionType as TypeofFunctionType,
 )
 from kirin.dialects.py.types.dialect import dialect as dialect
+from kirin.dialects.py.types.elem import (
+    PyAnyType as PyAnyType,
+    PyBottomType as PyBottomType,
+    PyClass as PyClass,
+    PyConst as PyConst,
+    PyGeneric as PyGeneric,
+    PyLiteral as PyLiteral,
+    PyType as PyType,
+    PyTypeVar as PyTypeVar,
+    PyUnion as PyUnion,
+    PyVararg as PyVararg,
+    hint2type as hint2type,
+    to_pytype as to_pytype,
+    widen_const as widen_const,
+)
